requirements.txt
feedparser>=6.0.2
click>=7.1.2
loguru>=0.5.3
python-dateutil>=2.8.1